+++ src/pages/survey-page/verify.tsx
@@ -1,16 +1,8 @@
 import React from 'react';
-import {types} from 'types';
-import {connect} from 'react-redux';
 import VisualTextCard from 'components/text-card/visual-text-card';
 import VisualInfoCard from '../../components/info-card/visual-info-card';
 
-function SurveyVerifyPage(props: {formConfig: types.SurveyConfig | undefined}) {
-    if (!props.formConfig) {
-        return <div />;
-    }
-
-    const config: types.SurveyConfig = props.formConfig;
-
+function SurveyVerifyPage() {
     return (
         <div className='w-full max-w-xl space-y-4'>
             <VisualTextCard title={'Verify now!'}>
@@ -22,10 +14,4 @@ function SurveyVerifyPage(props: {formConfig: types.SurveyConfig | undefined}) {
     );
 }
 
-const mapStateToProps = (state: types.ReduxState) => ({
-    formConfig: state.formConfig,
-});
-
-const mapDispatchToProps = (dispatch: any) => ({});
-
-export default connect(mapStateToProps, mapDispatchToProps)(SurveyVerifyPage);
+export default SurveyVerifyPage;
